Move download code into repo & service layer
