+++ .gitignore
@@ -148,6 +148,7 @@ venv/
 ENV/
 env.bak/
 venv.bak/
+.env.staging
 
 # Spyder project settings
 .spyderproject
